split SquiggleChart stories
