+++ packages/components/src/stories/SquiggleChart/Functions.stories.tsx
@@ -1,0 +1,34 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import { SquiggleChart } from "../../components/SquiggleChart.js";
+
+/**
+ * There are two types of functions that we can render:
+ * 1. number => number
+ * 2. number => dist
+ */
+const meta = {
+  component: SquiggleChart,
+} satisfies Meta<typeof SquiggleChart>;
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const FunctionToDistribution: Story = {
+  name: "Function to Distribution",
+  args: {
+    code: "foo(t) = normal(t,2)*normal(5,3); foo",
+  },
+};
+
+export const FunctionToNumber: Story = {
+  name: "Function to Number",
+  args: {
+    code: "foo(t) = t^2; foo",
+  },
+};
+
+export const Unrenderable: Story = {
+  args: {
+    code: "foo(x, y) = x + y; foo",
+  },
+};
